Refactor: add service layer, security, validation, pagination, and tests
- Add CompanyService/EmployeeService between controllers and repositories
- Upgrade Spring Boot 2.5.3 -> 2.7.18, remove unused commons-io
- Add Spring Security with form login and BCrypt auth (admin/admin)
- Add input validation (@Valid, @NotBlank, @Email) with error display
- Change entity IDs from int to Long, add JPA cascade on Company->Employee
- Fix GET-based deletes to use POST with CSRF protection and confirm dialogs
- Add edit functionality for both Company and Employee
- Add pagination support with Previous/Next navigation
- Add @ControllerAdvice error handling with custom error page
- Externalize DB credentials via Spring profiles (dev/prod)
- Add 26 unit and integration tests (service + controller layers, H2 for tests)

+++ src/main/java/com/example/companyemployeespring/repository/CompanyRepository.java
@@ -3,6 +3,5 @@ package com.example.companyemployeespring.repository;
 import com.example.companyemployeespring.model.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyRepository extends JpaRepository<Company, Integer> {
-
+public interface CompanyRepository extends JpaRepository<Company, Long> {
 }
